# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Cuadra API
  version: "1.0.0"
  description: |
    REST API del backend de Cuadra POS.
    Multi-tenant: todos los endpoints filtran por `ownerId` del JWT.
    Auth: Bearer JWT en header `Authorization`.
servers:
  - url: http://localhost:4000
    description: Desarrollo local
  - url: https://cuadra-backend.railway.app
    description: Producción (Railway)

# ─── SECURITY ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductVariant:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        stock: { type: integer }
        price_modifier: { type: number, nullable: true }
        sku: { type: string, nullable: true }
    ProductCreate:
      type: object
      required: [name, price]
      properties:
        name: { type: string }
        price: { type: number }
        cost_price: { type: number, nullable: true }
        stock: { type: integer, default: 0 }
        stock_by_location:
          type: object
          additionalProperties: { type: integer }
          nullable: true
        min_stock_alert: { type: integer, default: 0 }
        image_url: { type: string, nullable: true }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true }
        sub_category: { type: string, nullable: true }
        unit: { type: string, nullable: true }
        tags:
          type: array
          items: { type: string }
          nullable: true
        barcode: { type: string, nullable: true }
        location_id: { type: string, nullable: true }
        variants:
          type: array
          items: { $ref: "#/components/schemas/ProductVariant" }
          nullable: true
    ProductResponse:
      allOf:
        - $ref: "#/components/schemas/ProductCreate"
        - type: object
          properties:
            id: { type: string }
            owner_id: { type: string }
            active: { type: boolean }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    StockAdjustRequest:
      type: object
      required: [adjustment, reason]
      properties:
        adjustment: { type: integer }
        reason: { type: string, enum: [restock, waste, correction] }
        notes: { type: string, nullable: true }

    # ── Sales ───────────────────────────────────────────────────────────────
...
    ExpenseCreate:
      type: object
      required: [description, amount]
      properties:
        description: { type: string }
        amount: { type: number }
        category: { type: string, nullable: true }
        cashbox_id: { type: string, nullable: true }
        cashbox_name: { type: string, nullable: true }
        location_id: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        evidence_url: { type: string, nullable: true }
    ExpenseResponse:
      allOf:
        - $ref: "#/components/schemas/ExpenseCreate"
        - type: object
          properties:
            id: { type: string }
            owner_id: { type: string }
            created_at: { type: string, format: date-time }

    # ── Cashboxes ────────────────────────────────────────────────────────────
...
    LocationCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        address: { type: string, nullable: true }
        phone: { type: string, nullable: true }
    LocationResponse:
      allOf:
        - $ref: "#/components/schemas/LocationCreate"
        - type: object
          properties:
            id: { type: string }
            owner_id: { type: string }
            active: { type: boolean }

    # ── Suppliers ────────────────────────────────────────────────────────────
...
security:
  - bearerAuth: []

# ─── PATHS ────────────────────────────────────────────────────────────────────
paths:
...
  /products:
    get:
      summary: Listar productos activos
      tags: [products]
      parameters:
        - name: location_id
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProductResponse" }
          description: OK
    post:
      summary: Crear producto
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductResponse" }
          description: Producto creado
  /products/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Obtener producto por ID
      tags: [products]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductResponse" }
          description: OK
        "404":
          description: No encontrado
    patch:
      summary: Actualizar producto
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductCreate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductResponse" }
          description: OK
    delete:
      summary: Soft-delete producto
      tags: [products]
      responses:
        "204":
          description: Eliminado
  /products/{productId}/adjust-stock:
    post:
      summary: Ajustar stock (restock / merma / corrección)
      tags: [products]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StockAdjustRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductResponse" }
          description: Stock actualizado
        "400":
          description: Stock insuficiente

  # ── Sales ───────────────────────────────────────────────────────────────────
...
  /expenses:
    get:
      summary: Listar gastos
      tags: [expenses]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ExpenseResponse" }
          description: OK
    post:
      summary: Registrar gasto
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExpenseCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExpenseResponse" }
          description: Gasto creado
  /expenses/{expenseId}:
    delete:
      summary: Eliminar gasto
      tags: [expenses]
      parameters:
        - name: expenseId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Eliminado

  # ── Cashboxes ─────────────────────────────────────────────────────────────────
...
  /locations:
    get:
      summary: Listar sucursales
      tags: [locations]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LocationResponse" }
          description: OK
    post:
      summary: Crear sucursal (owner/admin)
      tags: [locations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LocationCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LocationResponse" }
          description: Sucursal creada
  /locations/{locationId}:
    parameters:
      - name: locationId
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Actualizar sucursal
      tags: [locations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LocationCreate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LocationResponse" }
          description: OK
    delete:
      summary: Desactivar sucursal (owner)
      tags: [locations]
      responses:
        "204":
          description: Desactivada

  # ── Suppliers ─────────────────────────────────────────────────────────────────
...